Moved tower place controller
